chore: remove /bin
