+++ src/main/java/com/aeon/biblioteca/entities/Autor.java
@@ -1,4 +1,4 @@
-package com.aeon.biblioteca.entityes;
+package com.aeon.biblioteca.entities;
 
 import java.util.Date;
