+++ src/main/java/com/aeon/biblioteca/entities/Livro.java
@@ -1,4 +1,4 @@
-package com.aeon.biblioteca.entityes;
+package com.aeon.biblioteca.entities;
 
 import java.util.Date;
