+++ src/main/java/com/aeon/biblioteca/entities/Pessoa.java
@@ -1,0 +1,5 @@
+package com.aeon.biblioteca.entities;
+
+public class Pessoa {
+
+}
