--- src/main/java/com/aeon/biblioteca/entityes/Pessoa.java
@@ -1,5 +1,0 @@
-package com.aeon.biblioteca.entityes;
-
-public class Pessoa {
-
-}
