+++ src/main/java/com/aeon/biblioteca/repositories/AutorLivroRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.aeon.biblioteca.entityes.AutorLivro;
+import com.aeon.biblioteca.entities.AutorLivro;
 
 @Repository
 public interface AutorLivroRepository extends JpaRepository<AutorLivro, Long>{
