+++ src/main/java/com/aeon/biblioteca/repositories/AutorRepository.java
@@ -2,7 +2,7 @@ package com.aeon.biblioteca.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.aeon.biblioteca.entityes.Autor;
+import com.aeon.biblioteca.entities.Autor;
 
 public interface AutorRepository extends JpaRepository<Autor, Long>{
