+++ src/main/java/com/aeon/biblioteca/repositories/CategoriaRepository.java
@@ -2,7 +2,7 @@ package com.aeon.biblioteca.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.aeon.biblioteca.entityes.Categoria;
+import com.aeon.biblioteca.entities.Categoria;
 
 public interface CategoriaRepository extends JpaRepository<Categoria, Long>{
