+++ src/main/java/com/aeon/biblioteca/repositories/EditoraRepository.java
@@ -2,7 +2,7 @@ package com.aeon.biblioteca.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.aeon.biblioteca.entityes.Editora;
+import com.aeon.biblioteca.entities.Editora;
 
 public interface EditoraRepository extends JpaRepository<Editora, Long>{
