+++ src/main/java/com/aeon/biblioteca/repositories/LivroCategoriaRepository.java
@@ -2,7 +2,7 @@ package com.aeon.biblioteca.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.aeon.biblioteca.entityes.LivroCategoria;
+import com.aeon.biblioteca.entities.LivroCategoria;
 
 public interface LivroCategoriaRepository extends JpaRepository<LivroCategoria, Long>{
