+++ src/main/java/com/aeon/biblioteca/repositories/LivroRepository.java
@@ -1,11 +1,9 @@
 package com.aeon.biblioteca.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-import com.aeon.biblioteca.entityes.Livro;
+import com.aeon.biblioteca.entities.Livro;
 
-@Repository
 public interface LivroRepository extends JpaRepository<Livro, Long> {
 	
 }
