+++ src/main/java/com/aeon/biblioteca/repositories/LivroTipoRepository.java
@@ -2,7 +2,7 @@ package com.aeon.biblioteca.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.aeon.biblioteca.entityes.LivroTipo;
+import com.aeon.biblioteca.entities.LivroTipo;
 
 public interface LivroTipoRepository extends JpaRepository<LivroTipo, Long> {
