+++ src/main/java/com/aeon/biblioteca/repositories/SituacaoRepository.java
@@ -2,7 +2,7 @@ package com.aeon.biblioteca.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.aeon.biblioteca.entityes.Situacao;
+import com.aeon.biblioteca.entities.Situacao;
 
 public interface SituacaoRepository extends JpaRepository<Situacao, Long>{
